# back/openapi.yaml
openapi: 3.1.0
info:
  title: VLR Backend
  version: 0.1.0
paths:
...
  /api/v1/attempts/:
    get:
      tags:
        - Попытки
      summary: Список попыток
      operationId: list_attempts_api_v1_attempts__get
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AttemptListItem"
                type: array
                title: Response List Attempts Api V1 Attempts  Get
      security:
        - HTTPBearer: []
    post:
      tags:
        - Попытки
      summary: Начать выполнение лабы
      operationId: create_attempt_api_v1_attempts__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptCreate"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
...
  /api/v1/attempts/{attempt_id}:
    delete:
      tags:
        - Попытки
      summary: Удалить попытку
      operationId: delete_attempt_api_v1_attempts__attempt_id__delete
      security:
        - HTTPBearer: []
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: integer
            title: Attempt Id
      responses:
        204:
          description: Successful Response
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/v1/components/:
    get:
      tags:
        - Компоненты
      summary: Мои составные компоненты
      operationId: list_components_api_v1_components__get
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ComponentResponse"
                type: array
                title: Response List Components Api V1 Components  Get
      security:
        - HTTPBearer: []
    post:
      tags:
        - Компоненты
      summary: Сохранить схему как компонент
      operationId: create_component_api_v1_components__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentCreate"
        required: true
      responses:
        201:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
  /api/v1/components/{component_id}:
    delete:
      tags:
        - Компоненты
      summary: Удалить компонент
      operationId: delete_component_api_v1_components__component_id__delete
      security:
        - HTTPBearer: []
      parameters:
        - name: component_id
          in: path
          required: true
          schema:
            type: integer
            title: Component Id
      responses:
        204:
          description: Successful Response
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/attempts/{attempt_id}/export.xls:
    get:
      tags:
        - Экспорт
      summary: Журнал замеров файлом Excel
      operationId: export_attempt_api_v1_attempts__attempt_id__export_xls_get
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: integer
            title: Attempt Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema: {}
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    AttemptCreate:
      properties:
        lab_id:
          type: integer
          title: Lab Id
        mode:
          type: string
          enum:
            - test
            - work
          title: Mode
      type: object
      required:
        - lab_id
        - mode
      title: AttemptCreate
...
    AttemptListItem:
      properties:
        id:
          type: integer
          title: Id
        student_id:
          type: integer
          title: Student Id
        lab_id:
          type: integer
          title: Lab Id
        mode:
          type: string
          title: Mode
        status:
          type: string
          title: Status
        result_data:
          anyOf:
            - $ref: "#/components/schemas/AttemptResultData"
            - additionalProperties: true
              type: object
            - type: null
          title: Result Data
        grade:
          anyOf:
            - type: integer
            - type: null
          title: Grade
        comment:
          anyOf:
            - type: string
            - type: null
          title: Comment
        started_at:
          type: string
          format: "date-time"
          title: Started At
        finished_at:
          anyOf:
            - type: string
              format: "date-time"
            - type: null
          title: Finished At
        lab_title:
          type: string
          title: Lab Title
        student_name:
          type: string
          title: Student Name
        class_number:
          anyOf:
            - type: integer
            - type: null
          title: Class Number
      type: object
      required:
        - id
        - student_id
        - lab_id
        - mode
        - status
        - started_at
        - lab_title
        - student_name
      title: AttemptListItem
    AttemptResponse:
      properties:
        id:
          type: integer
          title: Id
        student_id:
          type: integer
          title: Student Id
        lab_id:
          type: integer
          title: Lab Id
        mode:
          type: string
          title: Mode
        status:
          type: string
          title: Status
        result_data:
          anyOf:
            - $ref: "#/components/schemas/AttemptResultData"
            - additionalProperties: true
              type: object
            - type: null
          title: Result Data
        grade:
          anyOf:
            - type: integer
            - type: null
          title: Grade
        comment:
          anyOf:
            - type: string
            - type: null
          title: Comment
        started_at:
          type: string
          format: "date-time"
          title: Started At
        finished_at:
          anyOf:
            - type: string
              format: "date-time"
            - type: null
          title: Finished At
      type: object
      required:
        - id
        - student_id
        - lab_id
        - mode
        - status
        - started_at
      title: AttemptResponse
    AttemptResultData:
      properties:
        measurements:
          items:
            $ref: "#/components/schemas/Measurement"
          type: array
          title: Measurements
        answers:
          additionalProperties: true
          type: object
          title: Answers
        client_metrics:
          additionalProperties: true
          type: object
          title: Client Metrics
      type: object
      title: AttemptResultData
...
    ComponentCreate:
      properties:
        name:
          type: string
          title: Name
        graph:
          additionalProperties: true
          type: object
          title: Graph
      type: object
      required:
        - name
        - graph
      title: ComponentCreate
    ComponentResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        graph:
          additionalProperties: true
          type: object
          title: Graph
        created_at:
          type: string
          format: "date-time"
          title: Created At
      type: object
      required:
        - id
        - name
        - graph
        - created_at
      title: ComponentResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Measurement:
      properties:
        step:
          type: integer
          title: Step
        inputs:
          additionalProperties:
            type: number
          type: object
          title: Inputs
        outputs:
          additionalProperties:
            type: number
          type: object
          title: Outputs
      type: object
      required:
        - step
        - inputs
        - outputs
      title: Measurement
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer